Initial commit for COOL compiler
This simply implements the 4 basic phases of the cool compiler
in C++, without any significant optimization. Each phase is provided
as a separate project.

I am planning to significantly refactor this in the future to provide
a solid compiler for the language for my own experimentation.

For now, please accept the project as-is.